Feat: Line 거리 계산 구현

+++ src/main/java/challenge/coordinatescalculator/Position.java
@@ -1,19 +1,19 @@
 package challenge.coordinatescalculator;
 
 public class Position {
-    private int x;
-    private int y;
+    private double x;
+    private double y;
 
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 }
